disp: msm: dp: force link training for always connected DP
If always connected DP is enabled, the DP connection status will remain
connected. So after the DP is connected, link training will be skipped
in the hpd high handler, since the controller is alreay turned on. Then
the link training will be deferred to prepare stage, where by default
shallow training is applied. In case of the training failure, the link
training will not be retried according to the spec, lead to the display
can't be light up anymore.

The change will force the full link training procedure instead of shallow
mode in the prepare stage, if the always connected DP feature is enabled.

And for the case where the DP power is turned off and on from the user app
via the DRM ioctls, also force the full link training instead of shallow
mode, when always connected DP feature is enabled.

Additionally, for the link maintenance, add the link training retry for
always connected DP case.

Change-Id: I7e34fa411a7f680b1c41a81341e2a64b83e3a60f
Signed-off-by: Jim Wang <jimwan@codeaurora.org>
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>
Signed-off-by: Manting Yao <mantyao@qti.qualcomm.com>