disp: msm: dp: swap the connector has_tile and tile_group update order
Previously when parsing the DP connector tile information, has_tile flag
is cleared then check if the display is tile, and set has_tile flag and
update tile_group. However, if another thread is checking the connector
tile information, there might be racing condition that has_tile is set,
but tile_group is not updated yet.

Swap the updating order of the has_tile and tile_group:

1. If display has no tile information: clear has_tile in on shot
2. If display has tile information:
   2.1 Update tile_group first if it doesn't exist
   2.2 Set has_tile flag last

This will maintain the consistency of has_tile and tile_group. And
there would be no true->false->true transistion for has_tile flag,
when the function is callled but no tile information changes at the
display.

Change-Id: I882c623935017cefa0fab3c9cc65a9d559cc623a
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>