disp: msm: dp: keep link clock on for force connect mode
For force connect mode, user space expect can always render to the
display pipeline, regardless of the readiness of the sink device.
The requires the DP driver maintains the lclk and pclk always on
(except during link training and mode change commit), even the link
training is not completed with the sink device, either link training
is failed, or the HPD is disconnected.

So when link training failed for whatever reason, re-enable the
link clock as-is, and return link training success status to the
up layer to keep initializing the video stream.

Change-Id: I77c5d52263bbc72e25806d04ccda087447d66dc8
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>
(cherry picked from commit aa693f0471f661eaa72d4d85e3f862b85c1b59ac)