disp: msm: sde: signal cwb retire fence on cwb overflow
In the current SW, cwb overflow interrupt is not enabled for
mid and low tier targets due to which CWB retire fence is
not signaled, causing HWCBufferSyncHandler SyncWait timeouts
on fence. This change enables the cwb overflow interrupt
always to detect the overflow and signal the retire fence in
such cases.

Change-Id: Id6c47e20c92be104408e4c40c71ecb0b9f49a70e
Signed-off-by: Jayaprakash Madisetty <quic_jmadiset@quicinc.com>