disp: msm: dsi: add a new API to program slave PLL
In dual display cases where the PLL is shared across the displays,
some of the PLL registers for the slave display need to be programmed
separately because the master display is not aware of the slave when
it turns on.

Add a new API to allow programming the slave PLL registers separately.

Change-Id: Id07d2e4f2bf466f881239bfcd06a044c969879c1
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>