disp: msm: dsi: add a display manager to handle sync panels
Currently in dual dsi displays, each display is independent and
is unaware of the other one. In async mode this is fine as they
are mutually exclusive but in sync mode they share the PLL and
possibly other resources.

Add a dsi display manager module to manage the operations of
dual displays operating in sync mode.

Change-Id: I3fcad05c7a3df3e2c94edf4ed4086d5dfa4095b7
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>