disp: msm: dsi: ensure that the PHY timings are updated before enabling PHY
For sync panels, during boot up if the slave display turns on before the
master, display_set_mode for master would not have been called. This would
result in a failure during PHY enable of the master PHY, as the PHY timings
for it would have never been updated. This change ensures that the PHY timings
are updated before PHY is turned on.

Change-Id: Icb4341ccd8d3e7471e259ec148be3a747d63b1fd
Signed-off-by: Satya Rama Aditya Pinapala <psraditya30@codeaurora.org>
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>