disp: msm: dsi: Add support for 4 PLLs
Dual dpu instances are needed to support higher resolutions in VR
application. In dual dpu, 4 PLLs are supported.

In sync mode, PLL0 acts as master and other PLLs (PLL1, PLL2, PLL3)
act as slave. So, in this case there are 3 slave pll for master PLL.

In async mode, PLL0 acts as master of PLL1 and PLL2 acts as master
of PLL3. In this case there is 1 slave for each master PLL.

This change adds support for 4 PHY probe, clock registation for dpu1,
handling 3 slaves for master PLL and programming of slave PLLs in
sync mode.

Change-Id: Id613103c8da1ae3c7d7b42a55f460d97f8688e5a
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>