disp: msm-hyp: allocate gem objects for framebuffer of each plane
In msm-hyp driver we used to allocate only one framebuffer object
irrespective of the number of planes. But in video usecases we will
have more than one planes and used to result in atomic commit failures
as we were not able to allocate the framebuffer object for the second
plane.

This change is to allocate the gem objects to framebuffers
depending on number of planes.

Change-Id: I934cb5948477ec74a73f605185f1d34775b2c8a0
Signed-off-by: Ramya SR <quic_rsr@quicinc.com>