disp: msm: dsi: configure display manager for continuous splash
During continuous splash use case, the PLL and panel power are
already enabled, but the PLL resource refcount is zero and the
panel->powered flag is false. Due to this, the PLL and the panel
power do not get disabled during suspend. To avoid this, add a
vote on the PLL resource refcount and panel->powered in the
continuous splash use case.

Change-Id: If64feb6cf0e9c72638a8e84f3c9803a90c639d43
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>