disp: msm: dp: restart the DP session with normal mode for suspend to sim mode
If the DP driver suspend at simulation mode while the HPD is high, and resume
with HPD stay high as well, we end up with simulation mode. But assuming the
sink device has been restarted as well during the STR, so we need to restart
a new hotplug session with normal mode in this case.

Change-Id: Idb368e5bc6cc9335e74ed7a2118f7b332d49045b
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>
Signed-off-by: Manting Yao <mantyao@qti.qualcomm.com>