disp: msm: dp: add eDP panel notifier support
If panel notifier support is set, get panel phandle and add panel
to drm_panel. Notifier for this panel will be called on dpms update.

The feature is enabled via the 'qcom,panel-notifier-support' DT
property. When set, the driver looks up the panel node via the
'qcom,edp-default-panel' phandle and registers it with the DRM
panel framework so notifier consumers can find it via
of_drm_find_panel().

Change-Id: I6346eb09a551a993ec9050797d1ba9383491e612
Signed-off-by: Sai Srujana Oruganti <osaisruj@qti.qualcomm.com>