disp: msm: sde: unlock connection_mutex when wait pending commit
The global connection_mutex is used to protect all connectors' state,
and the routing chain of connector/encoder/CRTC. This mutex will be
locked when entering majority atomic_commit calls, and unlocked once
exit from the atomic_commit call. If one commit is delayed or the next
commit comes early for the same CRTC, the start_atomic for the second
commit has to wait for previous pending commit finished before submit
itself, meanwhile with the global connection_mutex is locked.

This will not be a hugh problem for single thread compositor system,
but will cause blockage of multiple compositor systems, or compositor
with multiple committing threads.

Such problem is seen in container environment, that if SurfaceFlinger
in Android container commit multiple frames in one Vsync period, or
delay a commit into next frame period, double with commit scheduled
for next frame, the second commit in the Vsync period will block
any commit from other containers, even on different connector/CRTC.

The solution is backoff the global connection_mutex when entering
start_atomic, where the waiting for previous pending commit_done is
performed, and lock the mutex again after the waiting. Meanwhile,
the mutex for each individual CRCT/plane/private object related
to the commit are still remain locked, to protect their atomic
states. If there is no need to wait for previous commit_done, then
this unlocking/relocking of connection_mutex can be skipped.

Since the connection_mutex is used to protect all connectors' atomic
state and the routing chain, unlocking it introduce a period that
they are not protected from multi-thread accessing. But as long as
there is no use case of multiple threads doing commit for the same
connector but different CRTC, it should be safe. Because if the
concurrent commits are performed on the same connector and same CRTC,
the CRTC object's atomic state mutex will still maintain the commits
are serialized, regardless the connection_mutex status.

In case of failure when relocking the connection_mutex, a retry and
bail out machanism is required to avoid freezing of the system. Even
DRM framework doesn't expect the atomic_commit stage to be failed,
but since before start_atomic we only prepare the planes (importing
framebuffer), and fences, no actual commit is done to the hardware,
so should be safe to bail out with EDEADLK error and let DRM framework
rollback all locks and resources and retry the commit from scratch in
a slower path.

Change-Id: I90a2522bb6ae22138ecebb0206891d2e719235c4
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>
(cherry picked from commit 2c68d14f0f32928e2edd07c2fa4d0f07e0e63a9c)