disp: msm: dp: support separated PLL for bond pclk
For pclk bonding DP mode, requires using a separated PLL to
generate the global pclk for all DP streams. This will avoid
the overshot of the deviation casued by opposite phase of link clock
drift coming from multiple DP PHYs when SSC is enabled.

To provide the PLL for the bonding pclk, need add following
properties into the devicetree:

- "bond_pixel_parent" and corresponding clock node to be added to
  "clock-names" and "clocks" node.

- New "qcom,dp-pclk-bond-pll" property points to the common DP PHY
  host the PLL, if the PLL is located in a DP/eDP PHY, instead of
  a generic clock source programmed by clock driver (dispcc).

- "dp_bond_phy", "dp_bond_pll", "dp_bond_ln_tx0", "dp_bond_ln_tx1"
  and corresponding register map to "reg-names" and "regs" node,
  if the PLL is located in a DP/eDP PHY.

- "core_ref_clk" and corresponding clock node to be added to
  "clock-names" and "clocks" node if the PLL relies on an external
   reference clock or regulator.

For the bond slave DP connector, mark it as "unknown" connect status
instead of "disconnected", to avoid the DRM framework wipe out the
EDID and display modes.

Change-Id: I4c6df7a6798215d6d75bea26549d3b05220c6b6f
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>