disp: msm: sde: add support for CWB + single LM partial update
In the current SW design for a CWB commit the need for 3D-Merge
and number of CWB mux which needed to tap the LMs are decided
based on number of mixers on crtc. When there is a partial update
commit in single LM these 3D-Merge and CWB mux active are used
leading to WB commit failures. This change properly checks whether
3D-Merge is needed in CWB path based on number of LMs in a partial
update commit.

Change-Id: I2c838a24ad3a259923f6b26934e681cb9a5829b6
Signed-off-by: Mahadevan <quic_mahap@quicinc.com>
Signed-off-by: Jinfeng Gu <quic_gjinfeng@quicinc.com>