disp: msm: sde: skip wait event for encoder not enabled successfully
If the encoder is not enabled successfully, usually downstream DP/DSI
isn't initialized successfulluy, the request for waiting for next event
shall return right aways, instead of waiting and timeout.

Fine tune the wait vsync/commit timeout based on the current display mode
with a small torlerance instead of hardcoding to worst case.

Change-Id: Iecb77e37b3968705d14e0930311eb7ce9d2f3828
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>