disp: msm: sde: add topology_name in crtc state
currently topology_name is defined as connector property and
when accessed by crtc in either check phase or commit phase,
the old state is always returned. This change copies the
topology_name to crtc state to simplify the searching and also
fix the old state used in check phase.Ported related changes for
the shd which is snapshot of commit a74e2ed.

Acked-by: akshaysk <akshaysk@qti.qualcomm.com>
Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>
Change-Id: Ie3f7a1372185e8909f07fea30b16daedf0d7d90a