disp: msm: sde: check enc_id in decreasing resource info
when the h/w blocks are reserved in the encoder's mode
set, the `num_lm` value  maintained in the sde resource manager's
available resources is decremented by the current number of
set encoders. The available resources should be decremented
for the newly added encoder only.

so, check if the encoder_id set in the hw blk is the same
as the id of the encoder during modeset. This should ensure that
we won't be counting the hw_blk more than once in decreasing
the aval_res count when the dymanic resource counter is refreshed.

Change-Id: Ice5b35cf4e35a3b850f346653384cddbf9bc93c2
Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>