disp: msm: dp: move GPIO interrupt handling to workqueue
When GPIO interrupt triggers, we need to flip the triggering edge.
Do this in IRQ itself is the best performance wise, but there is
corner case that during STR, if there is an interrupt triggered
during the early stages of the resuming, i.e. resume_noirq, that
will cause the GPIO irq function called before GPIO driver is fully
resumed. We will not get the context of the GPIO interrupt.

So we have to move the GPIO interrupt handling to workqueue,
sacrifice a bit performance, though not critical. Anyway there are
other hotplug state machine handling and timer setup steps need
to be performed in the the bottom half.

Also, solved an unbalanced enabled_irq warning. devm_request_irq
already enables the irq, so no need to manually call it again.

Change-Id: Ide83bfb1cd10c94412e9b45be95836c71773053a
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>
Signed-off-by: huiyang <quic_huiyang@quicinc.com>