sde: drm: fix the lm resource leak
The leakage fix resolves the LM not available issue for
the DP display ports.

along with this the logic for allocation and deallocation
of the LMs are enhanced to avoid allocating the resources
to the SHD, as SHD works on the copy of base display's resources.

Change-Id: I5a02747265286e5e8adbad6155be15c59f41b487
Signed-off-by: Eric Joseph <quic_erijos@quicinc.com>
(cherry picked from commit 62437713f609b6cd5fdeff8bdc47d5e38ee8098b)