disp: sde: change the log level for connector late register
dp connectors are usually created when they are added
in the device tree. When the DP MST is not enable,
meaning when the DP MST mode is not connected
there would not be any DP MST display which can be
enable and hence the late register function returns EINVAL.

The change update the log level from error to debug
to avoid dumping the logs when DP MST displays are not connected.

Change-Id: I26d336efefacbe32b1cf6f78cea271357e181727
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>