disp: msm-hyp: remove buffer channel and fix wrong message order
The change remove the usage of buffer channel from the display driver
as there is an optimization done in MM HAB driver to introduce
sync import and async unimport.

There is an another change done to fix the issue of wrong message
order where unexport is reaching hypervisor before backend unimport
due to the usage of sleep for hab_send_and_recv.

Change-Id: I647e07440e8b84fb8b2f0dacdb3e9b2bb8975711
Signed-off-by: Tushar Shivajirao Ghadge <quic_tghadge@quicinc.com>