msm: disp: dp: mark alt_mode_cfg_done in HPD recovery path
When HPD is mis-detected, the recovery function is called with correct
HPD status, but the alt_mode_cfg_done is not updated, lead to the
is_ready check failed later.

Update the alt_mode_cfg_done according to the reliable HPD
rerference (GPIO) as well, besides of the HPD status.

Change-Id: Ic61977d486b9763361992d0afe4cece40e742c19
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>