disp: msm: sde: check for connector type before dsi_display query
Code was trying to query dsi_display, without checking if the
connector is of DSI type. For DP type, code was not able to
proceed to respective connector pre_kickoff in case of eDP2, where
dsi_display was not available and flicker was seen in eDP2.

Added check for connector type before dsi_display query, and
shared display is now properly seen in eDP2 without flicker.

Change-Id: Ibad219f1270b27d03da3394fd0aab54ca8783a5a
Signed-off-by: Tushar Shivajirao Ghadge <quic_tghadge@quicinc.com>