msm: dp: bring down dp in suspend prepare
In the suspend path, there is a possible scenario
where DP stream is not enabled but aux is powered,
when suspend is requested on DP.

In this case, if DP is not enabled and powered on
i.e. aux is powered on then calling deinit directly
will not bring down the dp as init state is not set
and de-init will simply return without doing anything.

Additionally, if dp is initialized, ready but
not enabled, then we need to first unready
the display host to deinit the aux clocks before
we can call deinit on the display host.

So, to ensure that in this case the dp clocks
are bought down and they don't block suspend,
we need to handle this corner case in suspend
prepare.

Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>
Change-Id: I535fe7178862b9d63362db95f604de2baef2bba0