disp: msm: dp: disable DP host & aux in DPMST mode
During LPM, when suspend is triggered on DP MST setup,
device is not entering the suspend mode, because
DP host and aux is not turned off purposefully
to receive hotplug side band messages,
which is causing LPM TCs failure.

added a check to ensure DP host and aux is turned off
when unprepare is called from the suspend path.

Change-Id: I96ef56bbb42bd78a46b4d3793a4c5384f40736c8
Signed-off-by: Neha Kauser <quic_kauser@quicinc.com>