msm: dp: prevent dp deinit when its not connected
In suspend path, there is a possible scenario,
when DP stream is not enabled but aux is powered on,
when suspend is requested on DP.

In this case, when DP port itself is not connected,
then turning off DP would cause suspend/resume faiure.

Hence adding a condition before deiniting the DP to check
if port is connected or not.


Change-Id: Ic037c6ec656127e1a69936fbc9ef5b4b2e2fc0b4
Signed-off-by: Neha Kauser <quic_kauser@quicinc.com>
(cherry picked from commit 9595961e0a6d00737fa76e4436426089543910e1)