disp: msm-hyp: virtio: terminate panel name string with newline
panel name string was not terminated with new line,
which was causing invalid edid info at surfaceflinger.
we were populating 13 bytes(PANEL_NAME_LEN) string with panel name,
ex. panel name = DSI_1, which needs 5 bytes, and remaining 8 bytes
has value as 0x0.
at surfaceflinger we have panel name check which searches for '\n' and
verifies that all the bytes before '\n' are printable or not.
as we were not terminating panel name string with '\n', all 13 bytes
were checked and check was failing as 0x0 is not printable character,
which resulted in invalid edid info at SF.
added '\n' at the end of panel name string to separate printable and
non printable data.

Change-Id: Ifc82220a23968c17ee95666ea18c269490a916d5
Signed-off-by: Tushar Shivajirao Ghadge <quic_tghadge@quicinc.com>