disp: msm: sde: RM block changes for loopback encoder
During cac loopback usecases, we will have two encoders
(loopback and primary) tied to a single crtc. All the
resources needed for the first pass are handled by the
loopback encoder and all the resources for the second
pass are handled by the primary encoder. This change
add support to handle resource allocation for both the
encoders based on the current usecase. This change
also make sure that DSPP and DS blocks are reserved
in the first pass whenever cac loopback is enabled.

Change-Id: If129cced94beb5997ae0169a4c981cb47e0da9a2
Signed-off-by: Krishna Manikandan <quic_mkrishn@quicinc.com>
Signed-off-by: Sampurna Bolloju <quic_sampboll@quicinc.com>