disp: msm: sde: update frame timeout sequence during flush synchronisation
When HW flush sync is enabled, frames will be picked by both the cores
only if flush is issued on both. There might be some cases where
flush on one core is delayed or is not even issued by HAL.
Update timeout sequence to force async mode and wait for
additional timeout period to avoid frame timeout in such cases.

Change-Id: I29608430993acced87877bf0dc8a32b2aa539bff
Signed-off-by: Krishna Manikandan <quic_mkrishn@quicinc.com>