disp: msm: dsi: use qcom,dsi-select-sec-sync-clocks to check sync mode
Currently, in sync mode, qcom,dsi-select-sec-clocks property is set to
_mclk for secondary display.

If DPU0 is faulty and DPU1 is working, DPU1 will use the clock source
present in qcom,dsi-select-sec-clocks as it is the secondary display.
The value of qcom,dsi-select-sec-clocks needs to change from _mclk to
clk2 when DPU0 is faulty.

Therefore, add a new property for sync mode detection, i.e. qcom,dsi-
select-sec-sync-clocks, which will point to _mclk for secondary display.
So that the existing property "qcom,dsi-select-sec-clocks" can be used
as clock source for the secondary display when DPU1 is enabled to work
independently in Async mode and when DPU0 is faulty/disabled.

Change-Id: I8c29ff0fac3845aecb3661ecb26187c8ed2f9373
Signed-off-by: Ayushi Makhija <quic_amakhija@quicinc.com>