disp: msm: dp: fix RCU stall issue
On enabling skip-hpd with consolidate build,
RCU no preempt stall is detected on one of
the CPU.

with skip-hpd feature, HPD interrupts are
not expected hence not addressed, but on
enabling skip-hpd there are continuous
interrupts are triggered related to HPD.

as there are no ISR registered to address
these interrupts, one of the CPU is busy
handling these interrupts and not preempted
for the long time leads to RCU stall.

while configuring the HPD at the init time,
skip-hpd is not considered, issue is fixed
with not configuring HPD when skip-hpd is
added in the dtsi.

Change-Id: Ieafe96fa344c36d5e7b1271065c10f013fd5beb0
Signed-off-by: Neha Kauser <quic_kauser@quicinc.com>