disp: msm: use rsc index instead of dpu id for rsc checks
In the current logic, dpu index is used to fetch the status
of the rsc block binded to the corresponding Dpu. But the
dpu id and rsc index wont be equal for cases where probe
of one core is delayed or one core is disabled in SW. This
change adds necessary support to use the correct index while
trying to fetch the rsc block of each Dpu.

Change-Id: Ie4a540717f127807b84b7e33bf2dff4e9e2a9a74
Signed-off-by: Krishna Manikandan <quic_mkrishn@quicinc.com>