disp: msm: shd: make base display plane accessible
Update the possible crtc mask for primary plane.

If the primary plane allocated to the base display
is made dummy and cannot be used by a crtc, this
takes away one possible plane to stage the incoming
layers on.

In presentation mode, i.e. the primary display is
rendering while the secondary display is not mirroring
the content, if all the pipes are consumed for the
primary display, and a new application started on
the secondary display we fail to allocate new pipes,
unless a hot plug is forced.

By setting possible crtcs for plane allocated to the
shared display base, we get improved performance in
such corner cases. We are still allocating a primary
plane for the shared display base interface, but no
longer making it a dummy.

Change-Id: I5abac5629c9b5464f1d119972d15565dbfb51cb7
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>
Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>
(cherry picked from commit a5b8cfeb54ad18960ef2fd4fc5289c18b5b6b792)