disp: msm: sde: add TUI support for VHM
PVM programs the hardware to treat the panel as a VHM panel, but
TVM is unaware of that and thinks that the panel is actually legacy
video mode. Due to this discrepancy, TVM must set fence SW override
for all frames when dealing with the VHM panel.

Disables AVR when PVM hands off hardware to TVM, and sets fence
SW override for all frames within TVM for VHM panel usecase.

Change-Id: I9542f2706dde6ad37fb3068cad00df9b71e41e67
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>