msm/dp: protect edid and hpd buffers with mutex
The EDID and HPD buffers are allocated and deallocated
in debug module while running in simulation mode. There
can be a race condition if multiple scripts are run and
result in double free of the buffers.

Protect the buffer allocation and deallocation functionalities
with a mutex so that such race conditions can be avoided.

Change-Id: Ia203a2eb582d341d7a19f8ede3313ac7287445dd
Signed-off-by: Fan Zhang <quic_fanzhan@quicinc.com>