disp: msm: dsi: fix dynamic bit clock for 14nm pll
Due to wrong n1 div value, the vco rate calculated is
incorrect and dynamic bit clk pll code is unable to find
the correct vco rate. So update the div value from 4 to 3
in the frequency table. As same table is used for other
targets, for now, add arch specific checks to select
correct table.

Change-Id: I1374eaffede764fcda4f47bb7173368e493a31de
Signed-off-by: quic_swarbu <quic_swarbu@quicinc.com>
Signed-off-by: Venkatesh Kotha <quic_venkkoth@quicinc.com>