disp: msm: sde: enable continuous smooth dimming
When smooth dimming for N frames is in progress
and a new backlight update is received, initiate
smooth dimming from the current backlight value
instead of switching to the new value.

For example:
- Initial backlight value: 100
- New backlight value: 1100 with 10 smooth dimming frames
- Expected dimming sequence: 200, 300, 400, ... 1000

If a new backlight value of 1400 is set when smooth
dimming is at 400, the dimming sequence should
adjust to: 500, 600, ... 1400.

Change-Id: Idd80af6c459f6f0d76d8a0f6d275cec950c6e0dc
Signed-off-by: Raviteja Tamatam <quic_travitej@quicinc.com>