disp: msm: sde: handle multiple BL updates during off work
In certain corner cases, scheduling delays may cause
the smooth dimming backlight incremental update to
coincide with the backlight turning off. Concurrently,
if a new backlight update occurs, it can sometimes lead
to a deadlock. To prevent this, implement an early return
from the smooth dimming backlight incremental update
when such a scenario is detected.

Change-Id: I15e996e7329ed6dcf03a81695069b3db783e613b
Signed-off-by: Raviteja Tamatam <quic_travitej@quicinc.com>