disp: msm: dsi: toggle TE GPIO to avoid idle delay
Hardware tracks TE level internally using the last seen edge in the
signal, and the default state is low. If TE is high when MDSS is
powered on, then until the next falling edge, hardware will think
that TE is low when it's actually high.

When connecting TE, toggle its GPIO function briefly to refresh
hardware's state tracking.

Change-Id: I8bf7b37a611052cde7dd615d420ab34ba87dcd53
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>
Signed-off-by: Lei Chen <quic_chenlei@quicinc.com>