disp: msm: sde: detect ESD on VHM using TE
VHM panel uses TE to signal to AP when it's currently busy updating
itself from GRAM, or receiving panel data from AP. If TE remains low
for longer than expected, this may indicate that an ESD event occurred
to the panel.

This change adds logic to keep track of TE deasserts as a response
to vblanks, and reports panel as dead when TE is not keeping up
with vblanks.

Change-Id: I3e252fa8a30e74e6d3b92159911f1fda810e1ecd
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>