disp: msm: dsi: add support to set esync clk rate
According to HPG, In pclk mode, esync_clk should be sourced
from the DSI PLL and the clock dividers should be configured
in the same way as pclk to provide same frequency as pclk.

Current DSI driver is not calling the clk_set_rate() API
explicitly for esync clock due to which MND is defaulting
to 1/1. This MND configuration works for 8 bit panels but
it results in mismatch between pclk and esync_clk rates
for 10 bit panels in which case MND should be configured to 2/3.

Add support to configure and set esync clock rate during
esync clock enablement and dynamic clock switch.

Change-Id: If1d50c1d915fe96bef98cb42cbdcf393bfc7eb2f
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>