disp: msm-hyp: fix one black screen issue when color mode changes.
When color mode changes from sRGB to P3, the blend_op will change
from PREMULTIPLIED to COVERAGE in SDM, but currently the COVERAGE
is set to OPAQUE as default in wfd kms driver which will
cause the bottom layer is masked by top layer unintentionally,
if the top layer is black, the whole panel will be black screen.

So, add COVERAGE of blend_op support to make the bottom layer could
be showed accordingly in this case.

Change-Id: Iec349fbd499846f71f3224c3654eac815e955a5e
Signed-off-by: jiabwu <quic_jiabwu@quicinc.com>