disp: msm: sde: reverse order encoder_prepare_for_kickoff
In the VHM panel, for idle-pc exit commit, if CWB is also enabled,
first CWB programming for that particular CTL path is done,
and later the primary encoder interface timing engine setup
is done, which clears the existing configuration leading to WB
commit failures. This change helps in proper programming in
such cases.

Change-Id: I8b6d17628e4362881d023968cbb55a29efa2372e
Signed-off-by: Mahadevan <quic_mahap@quicinc.com>