disp: msm: dsi: fix dynamic clock switch for 10-bit panels
The pclk_rate is not being updated in the PLL structure during
dynamic clock switch. As a result an icorrect pll->pclk_rate is
being returned to the clock driver which is causing MND
programming to always be set to 1/1. This works for an 8-bit
panel where the MND is 1/1. However, it causes DSI FIFO overflow
and underflow errors for other MND configurations. For example,
a 10-bit panel with an MND of 2/3 can encounter errors following
a dynamic clock switch.

Change-Id: I0214292803ddcea5437f36a9e7b8522f3bb81aab
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>