disp: msm: sde: disable timing engine only when enabled
In scenarios where there is no frame after an early wake-up
or when the system is idle followed by a suspend in the VHM
panel, the timing generator is already disabled. Re-disabling
it in this state leads to a timeout issue. To prevent this,
implement a check to ensure the timing engine is not disabled
again if it is already in a disabled state.

Change-Id: Ib22469bf9b0a911e172eb136a643505a6c0ff4b0
Signed-off-by: Raviteja Tamatam <quic_travitej@quicinc.com>