disp: msm: dp: Add MST check before parking stream 1 clock
When MST is not enabled, the pixel1_clk_rcg clk_get is not done.
During DP simulation, dp_power_park_clocks gets called if USB gets
disconnected where dp_power_park_module gets called for DP_STREAM1_PM
even though MST is disabled. This leads to failure in parking the
stream 1 clock and thereby link clock also does not get parked as it
returns early due to the failure.

This change adds a check for mst to avoid this issue.

Change-Id: I9d7f089f1bbd6962b19e64bcf56967a5555e5e95
Signed-off-by: Amit Singh <quic_amitsi@quicinc.com>