disp: msm: sde: update current backlight with requested backlight
When smooth dimming is enabled, the previous and current
backlight levels are used to calculate the backlight update
per frame over a set number of frames. If smooth dimming
is disabled before the backlight update of the last frame
is completed, the current and previous backlight levels retain
their last updated values. If there is backlight update request
for level (x) without smooth diming, and later new backlight
level request (y) with smooth diming, the backlight per frame
should be calculated within the range from (x) to (y). However,
since the previous and current backlight levels hold old
values, this calculation can be incorrect, leading to
flickering when smooth dimming is dynamically enabled and disabled.

To prevent this, update the current and previous backlight
levels to the configured backlight value when smooth dimming
is disabled.

Change-Id: Ic52b655d06373719faba24cd95c3ad2ea37ed86f
Signed-off-by: Abhinav Saurabh <quic_abhisaur@quicinc.com>