disp: msm: dsi: move esync_enabled variable from timing to host_config
There is a race condition when accessing the esync_enabled variable,
which is specific to panels that support esync. This variable is not
part of the drm_mode structure, and is later updated from the
host_config data. During set_mode, while the video engine is being
configured to the adjusted mode, the value of esync_enabled remains
0 until it is updated after setting up the video timing engine. If
this variable is accessed during this interim period, it can lead
to a race condition. This variable is used checked while updating ctrl
state. If another thread updates the ctrl_state during this intrim
time, it can result in ctrl_check_state failure.

To avoid this, refactor the timing structure by moving esync_enabled
variable from the timing structure to the host_config structure and
access this variable accordingly. This ensures that the variable is
not reset during video engine timing setup.

Change-Id: I0e534a48b5b01fc57d74a6441d81c62519e877c2
Signed-off-by: Anand Tarakh <quic_atarakh@quicinc.com>