disp: msm: dsi: load DSI prop only if selected
Currently, DSI prop is always attempted to be loaded when no panel
is selected. This results in a wait for firmware to be loaded, which
can be avoided when DSI prop isn't being used.

This change attempts to load firmware only if UEFI explicitly tells
driver to use DSI prop.

Change-Id: I53f1bff8ff1596c1eddf35b05c243378379f69b2
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>
Signed-off-by: Vaishali Gupta <quic_vai@quicinc.com>