disp: msm: dp: fix slice calculation for 8k60
This change adds a special case for slice calculation for 8k60 format.
As per spec, for the case 1360 < ppr <= 3200, the slice count of 8 is
selected assuming a peak throughput support of:

 340 MP/s for ppr < 2700 and
 400 MP/s for ppr > 2700

But an 8k60 sink can support a peak throughput of 550 MP/s to allow
a 4 slice configuration. So in this special case, the driver will
choose 4 slices instead of 8.
This change also sets the correct pixels per clock value for TU
calculation.

Change-Id: Icc8d5e32c069237367c64441585e80b5e877ada3
Signed-off-by: Rajkumar Subbiah <quic_rsubbia@quicinc.com>