disp: msm: sde: fix check in hw_fence_extended_wait
Fence_ready bit will always be set in command mode as this captures
internal fence_ready flop in DPU HW only during VSYNC, or when the DPU
HW successfully processes the commit.

Thus, remove this check and fix check for input sw signals. With this
change, we correctly assume that planes without input fences are
"signaled" instead of considering this as unsignaled case.

Change-Id: I2c8449e789032ea0362c6c7b69b33c1082be750f
Signed-off-by: Grace An <quic_gracan@quicinc.com>