disp: msm: dsi: avoid restoring bit clk & front porches during set mode
In video mode, if there are two consecutive commits, with a DFPS change
request in the Nth commit and a dynamic clock change request in the
N+1th commit, the N+1th commit is processed before the Nth commit is
served and set into DSI. During the mode validation of the N+1th commit,
the porches are updated according to the requested dynamic clock.
If bridge_set_mode and bridge_pre_enable of the Nth commit occur after
bridge_mode_fixup of the N+1th commit, the porches and clock restored
for the Nth commit will be disrupted by the N+1th commit. This discrepancy
will cause different programming in INTF and DSI, for the Nth commit
leading to overflow and underflow.

Since restoring the bit clock is already handled during bridge_mode_fixup,
it should be avoided during bridge_mode_set and bridge_pre_enable.

Change-Id: Ieecb0020a77f5e082a8b9da0ecf461acdbe89e0c
Signed-off-by: Anand Tarakh <quic_atarakh@quicinc.com>