disp: msm: dsi: read esync_clk with clk_master_idx of display
In case of dual display, an incorrect index is being used to
read the esync clock for the secondary display. This results
in 0 esync clock value, causing the backup esync configuration
to be incorrect and leading to flickering during idle PC exit.

To resolve this, use the clk_master_idx of the display
to correctly read back the esync clock as clock manager
differs for each display.

Change-Id: I452dc26406e57559e08f254b04882d16f042445b
Signed-off-by: Anand Tarakh <quic_atarakh@quicinc.com>