disp: msm: sde: avoid disabling infinite mode in the IPC exit and VM release commit
In some corner case scenarios, the idle PC exit and VM release
can occur in the same commit. Infinite mode is being enabled as
part of the VM release, but it is getting disabled again due to
the timing engine reprogramming during the IPC exit sequence,
which is causing a noisy screen. This change ensures that infinite
mode is not disabled. Additionally, it prevents re-enabling
infinite mode if it is already enabled, which helps avoid the MDP
INTF from hanging.

Change-Id: I9b258c1f927a3bcd54a643f1c7a83d7506842dde
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>