disp: msm: sde: select appropriate vsync source for VHM panel
In a VHM/MRR dual display setup, vblank timeout issue occurred
during the secondary display resume commit, while the primary
display remained off. This was due to the secondary display
interface block not being initialized with the appropriate vsync
source. This change ensures the selection of the relevant vsync
source for both the primary and secondary panels.

Change-Id: I328cf71c499a32911790d1c710706e009bfa1f0a
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>