disp: msm: sde: queue commit in the corresponding CRTC thread
In a dual display usecases, when the primary display is off
and the secondary display is on, if there is hw composer kill
usecase, drm atomic state is populated with both the crtc's
active set to false. With primary display in off state,
commit with the secondary display's active to false state will
be queued in the primary display's thread. This can lead to a
situation where rc_idle (running in the secondary thread) and
the secondary disable commit (running in the primary thread) might
race with each other. To prevent this, queue the commit to
the associated display thread.

Change-Id: Id64ad2f0986113c28dea1a5adbf9029950be3110
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>