disp: msm: dsi: refactor dsi_display_send_pre_commit_cmd() API
Currently dsi_display_pre_commit API is being used as a
connector API for command transfer which should ideally be
used only to perform pre commit DSI operations.

Refactor dsi_display_send_pre_commit_cmd() API to be used as
a connector interface to process DCS command bitmask so as to
send multiple command sets in a batch.

This refactor is a no-op change.

Change-Id: I1c50eee9bf5c511b7b88feb5abf6f45d15c0d268
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>