display: use of_find_node_by_name to get virtio/wfd node
Use "qcom,transport" property from sde-kms-hyp node, if
available, to get the virtio/wfd node.
Currently, this is needed in case of sdp8 in nord due
to phandle mismatch issue in sdp8.

If this property is not set, then use the existing
approach to use get virtio/wfd node using phandle.

Change-Id: Ic48c965e92b0d720cc7880ca7df40166579e097e
Signed-off-by: Shrikara B <quic_shrikara@quicinc.com>
Signed-off-by: Devi Mani <quic_dts@quicinc.com>