msm: sde: get ROI info based on DS enablement status
RC, Demura and SPR are post destination scaler. ROIs passed to
these blocks depend on enablement of DS. This change gets ROIs
based on DS enablement. ROIs are from connector state if DS
is enabled and from crtc state if DS is not enabled.

This change also removes cached rois as cached ROIs are not
used in determining PU processing.

Change-Id: I4fd77940f7aaeaaad9880f3a5a55145f279fc04d
Signed-off-by: Anjaneya Prasad Musunuri <quic_aprasad@quicinc.com>