disp: msm: sde: Fix plane attach frame to global flush
In the previous code, the first frame of plane attachment could enter
the logic for enabling SSPP local flush twice.
On the second entry, SSPP local flush would be forcibly enabled.

Changed the configuration from being controlled by the number of entries
to setting the flush type for the relevant plane when the CRTC selects
the flush type.

Change-Id: Id2860dd08283b449435a4a55895a007b3412003e
Signed-off-by: ruoyyang <quic_ruoyyang@quicinc.com>