disp: msm: dp: Remove REP_CONFIG_VERSION timeout
Addressing a thread scheduling issue that occurs when
the CPU is overloaded. This issue prevents the
threads from waking up and completing the REP_CONFIG_VERSION
update in a timely manner, leading to a timeout for the
REP_CONFIG_VERSION update.

To resolve this, the timeout for REP_CONFIG_VERSION has been
removed, and the RX_INFO is now cached in a local variable to
avoid any possibility of it being overwritten by the subsequent
REP_SEND_ACK message.

When the Repeater is HDCP 2.2 supported but Downstream device
isn't HDCP 2.2 supported:
/sys/devices/virtual/hdcp/msm_hdcp # cat hdcp_version
1

When the Repeater and the Downstream device is HDCP 2.2 supported:
/sys/devices/virtual/hdcp/msm_hdcp # cat hdcp_version
2

When the DP cable is connected directly to HDCP 2.2 supported display:
/sys/devices/virtual/hdcp/msm_hdcp # cat hdcp_version
2
.

Change-Id: I46bb72eab416e94a45ebc66d313f4de519a69017
Signed-off-by: Shubham Dhiman <shubhamd@qti.qualcomm.com>
(cherry picked from commit 85d7ca7a04dd904c444b1313256cb732b24a13c9)