display-driver: Fix SSR regression
SSR sequence is broken due to introduction of crtc dither property
payload buffer allocations and mapping to firmware which are not handled
as part of ssr sequence to deallocate when ssr starts, and re allocate
after completion of ssr sequence. This is causing iommu map failures
for the unfreed memory block allocated as part of initial hfi crtc inits.
Fixed this issue.

Change-Id: I1ec42d5ee11800ff71fdccde14c1325a8d271a43
Signed-off-by: Akshata Sahukar <asahukar@qti.qualcomm.com>