disp: msm: hfi: increase priority of HFI callback event thread
Add a dedicated worker to adjust the scheduling priority
of the HFI callback event thread. This ensures the thread
runs with higher priority, improving responsiveness and
reducing latency in handling HFI events. This change helps
in scenarios where WB and Primary Display are running
concurrently, as this helps to avoid timeout waiting for
the interrupt.

Change-Id: I1e70c9167a1a9491c32d25d0673527cc35989e2c
Signed-off-by: Govind R Sekhar <quic_gsekhar@quicinc.com>