disp: msm: dp: Reset HDCP version during suspend
Issue:
During the Suspend-to-RAM (STR) resume process, where a race condition occurs
due to the delayed execution of the dp_display_enable call. As a result,
dp_display_connect_work is triggered prematurely, leading to the invocation of
dp_display_clean. This function sends an HDCP_2X_CMD_STOP command to the secure
environment (QSEECOM). However, since the user space is still frozen at this
stage of resume—while the kernel is active—the call to QSEECOM gets blocked,
ultimately causing the system to hang.

Fix:
This change ensures that the HDCP version is reset during system suspend,
which helps in preventing operational failures related to HDCP handling
upon resume.
HDCP transactions are bypassed when the HDCP version is reset. For example,
in the observed issue, dp_display_clean() internally invokes hdcp_off,
which depends on the result of dp_display_is_hdcp_enabled(),
a check that is governed by the current HDCP version

Acked-by: Darshan N <darshann@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>
Change-Id: I690874c07d19203969e175f2686c88db198feb2d