disp: msm: sde: release CWB resources after primary flush
In the following scenario, resource conflicts can occur:
 1. A commit disables both the Primary Display and
    Concurrent WriteBack (CWB).
 2. The Primary Display is using CWB resources (PP0 and PP1).
 3. During resource manager (RM) release, if the primary flush
    hasn't been picked up, the secondary display (which intends
    to use CWB) may begin using the MUX and PP resources prematurely.
 4. This leads to CWB commit failures due to resource contention.

To address this, the RM release is deferred until the primary commit
is completed (wait_for_commit_done) if the primary display is active
and CWB is disabled. If both primary and CWB are disabled, the
release is triggered on virt_disable of the primary, after its commit
is picked up.

Change-Id: I93c51f439029068c00e9e1dd67f2e5c9645cad98
Signed-off-by: Mahadevan P <mahap@qti.qualcomm.com>