disp: msm: dp: remove display panel object from DRM
Remove DP panel object shared to DRM framework. DP Panel
is local DP module and should be used within DP.

This DP panel instance is specific to HWIO architecture.
This instance is not applicable to DCP. As DRM architecture
is common for HWIO and DCP, local modules need to be accessed
only within the module.

Use an integer to represent panel id instead. Locally, DP can
maintain the mapping of DP panel and id.

Change-Id: If5ab957b45e3429e94b8ff414777655dc61c2647
Signed-off-by: Ajay Singh Parmar <quic_aparmar@quicinc.com>