disp: msm: sde: decrease pending kickoff count when skip wait vblank
In the SHD use case, shared display and base display use the same
physics display, if the base display disabled, it would stop the
irqs and clks, shared display wait for vblank will timeout, for
this use case, shared display skip wait for vblank. But the pending
kickoff count only increases and not decrease, that causes the
pending kickoff count always increases, if the count more than
three and the base display re-enabled, the next commit need to wait
for vblank and will timeout 84ms.

Change-Id: I274e8c34c33e3fa798ac55c34b77e67b687e3b13
Signed-off-by: Xiaoliang Wang <xiaolwan@qti.qualcomm.com>